project-3/user-management-service/pyproject.toml
[project]
name = "user-management-service"
version = "0.1.0"
description = "User Management Service API"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "fastapi~=0.110.0",
    "uvicorn~=0.27.0",
    "sqlalchemy~=2.0",
    "alembic~=1.13.0",
    "passlib[bcrypt]~=1.7.4",
    "python-jose~=3.3.0",
    "pydantic-settings~=2.1.0",
    "pydantic[email]",
    "psycopg2-binary~=2.9.9", # PostgreSQL adapter
    "ruff~=0.3.0",
    "pytest",
    "pytest-cov",
    "testcontainers",
    "factory-boy",
    "httpx",
    "python-multipart",
    "sqlalchemy-utils",
    "pytest-mock",
    "pytest-asyncio",
    "freezegun",
]
requires-python = ">=3.11"
license = { text = "MIT" }

[tool.uv]

[tool.ruff]
line-length = 120
select = ["E", "F", "W", "I", "B", "C4", "N", "UP", "YTT", "ANN", "ASYNC", "S", "BLE", "A", "COM", "D", "ERA", "ISC", "ICN", "PIE", "PT", "Q", "RSE", "RET", "SLF", "TCH", "TID", "INT", "ARG", "PERF", "FURB", "RUF"]
ignore = ["D100", "D104", "D105", "D107", "ANN101", "ANN002", "ANN003", "ANN401"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["S101", "D", "S106"]

[tool.ruff.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
pythonpath = [".", "app"]
addopts = "--strict-markers --strict-config"
asyncio_mode = "auto"
